removed some migrations

+++ db/migrate/20170225014200_add_eps_columns_to_stocks.rb
@@ -1,7 +1,7 @@
 class AddEpsColumnsToStocks < ActiveRecord::Migration
   def change
-    add_column :stocks, :eps_curr_year, :decimal
-    add_column :stocks, :eps_last_year, :decimal
-    add_column :stocks, :eps_last_2_year, :decimal
+    add_column :stocks, :eps_curr_year, :float
+    add_column :stocks, :eps_last_year, :float
+    add_column :stocks, :eps_last_2_year, :float
   end
 end
